psr3: handle non-string attribute keys (#297)
non-string attribute keys cause an error, and stop log messages from being exported. cast the numeric keys
to a string to ensure they make it through alive. update tests, skipping cake because it does weird stuff
when presented with a numeric key which completely changes the structure of the incoming context